fix: 🐛 light svg on tablet and mobile devices and try to fix
laggy animation on tablet and mobile

✅ Closes: #63 #62
