+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "iSunFA",
-  "version": "0.1.1+3",
+  "version": "0.1.1+4",
   "private": false,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "next lint && eslint --fix --ext .js,.jsx,.tx,.tsx .",
     "test": "jest --coverage",
     "check-format": "prettier --ignore-path .gitignore --list-different \"**/*.+(js|jsx|ts|tsx|json)\"",
     "format": "prettier --ignore-path .gitignore --write \"**/*.+(js|jsx|ts|tsx|json)\"",
