+++ src/constants/display.ts
@@ -1,0 +1,1 @@
+export const IS_BUTTON_DISABLED_TEMP = true;
